+++ .github/workflows/python_tests_checks.yml
@@ -1,6 +1,6 @@
 name: Checks on the Python tests
 
-# This workflow performs some checks on the Python client used by the Boilerplate tests
+# This workflow performs some checks on the Python client used by the Ironfish tests
 # It is there to help us maintain a level of quality in our codebase and does not have to be kept on forked
 # applications.
 
@@ -16,7 +16,7 @@ on:
 jobs:
 
   lint:
-    name: Boilerplate client linting
+    name: Ironfish client linting
     runs-on: ubuntu-latest
     steps:
     - name: Clone
