+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "app-boilerplate-rust"
+name = "ledger-ironfish"
 version = "1.5.0"
 authors = ["yhql", "agrojean-ledger"]
 edition = "2021"
@@ -11,6 +11,9 @@ serde = {version="1.0.192", default_features = false, features = ["derive"]}
 serde-json-core = { git = "https://github.com/rust-embedded-community/serde-json-core"}
 hex = { version = "0.4.3", default-features = false, features = ["serde", "alloc"] }
 numtoa = "0.2.4"
+
+ironfish-frost = { git = "https://github.com/iron-fish/ironfish-frost.git", branch = "no-std", default-features = false, features = ["dkg"]}
+getrandom = { version = "0.2", features = ["custom"] }
 
 [profile.release]
 opt-level = 'z'
@@ -24,16 +27,21 @@ pending_review_screen = []
 curve = ["secp256k1"]
 flags = "0"
 path = ["44'/1'"]
-name = "Rust Boilerplate"
+name = "Ironfish"
 
 [package.metadata.ledger.nanox]
-icon = "crab_14x14.gif"
+icon = "nanox_icon.gif"
 
 [package.metadata.ledger.nanosplus]
-icon = "crab_14x14.gif"
+icon = "nanox_icon.gif"
 
 [package.metadata.ledger.stax]
-icon = "crab_32x32.gif"
+icon = "stax_icon.gif"
 
 [package.metadata.ledger.flex]
-icon = "crab_40x40.gif"
+icon = "stax_icon.gif"
+
+# TODO: This change shouldn't be necessary, the ledger targets clearly define atomics as only supporting 32 not 64 (as seen in /opt/rustup/toolchains/1.75.0-aarch64-unknown-linux-musl/lib/rustlib/nanosplus/target.json)
+# solve why this is happening rather than using modified radium
+[patch.crates-io]
+radium = { git = "https://github.com/iron-fish/radium", rev = "674c8faf1e74f931a58671f70586e6435353e9b6" }
