+++ README.md
@@ -1,4 +1,4 @@
-# Ledger Rust Boilerplate Application
+# Ledger Rust Ironfish Application
 
 ![Rule enforcer](https://github.com/LedgerHQ/app-boilerplate-rust/actions/workflows/guidelines_enforcer.yml/badge.svg) ![Build and tests](https://github.com/LedgerHQ/app-boilerplate-rust/actions/workflows/build_and_functional_tests.yml/badge.svg)
