+++ tests/application_client/boilerplate_command_sender.py
@@ -51,7 +51,7 @@ def split_message(message: bytes, max_size: int) -> List[bytes]:
     return [message[x:x + max_size] for x in range(0, len(message), max_size)]
 
 
-class BoilerplateCommandSender:
+class IronfishCommandSender:
     def __init__(self, backend: BackendInterface) -> None:
         self.backend = backend
