+++ tests/test_appname_cmd.py
@@ -1,12 +1,12 @@
-from application_client.boilerplate_command_sender import BoilerplateCommandSender
-from application_client.boilerplate_response_unpacker import unpack_get_app_name_response
+from application_client.ironfish_command_sender import IronfishCommandSender
+from application_client.ironfish_response_unpacker import unpack_get_app_name_response
 
 
 # In this test we check that the GET_APP_NAME replies the application name
 def test_app_name(backend):
     # Use the app interface instead of raw interface
-    client = BoilerplateCommandSender(backend)
+    client = IronfishCommandSender(backend)
     # Send the GET_APP_NAME instruction to the app
     response = client.get_app_name()
     # Assert that we have received the correct appname
-    assert unpack_get_app_name_response(response.data) == "app-boilerplate-rust"
+    assert unpack_get_app_name_response(response.data) == "ledger-ironfish"
