+++ tests/test_error_cmd.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ragger.error import ExceptionRAPDU
-from application_client.boilerplate_command_sender import CLA, InsType, P1, P2, Errors
+from application_client.ironfish_command_sender import CLA, InsType, P1, P2, Errors
 
 
 # Ensure the app returns an error when a bad CLA is used
